15 Aug 11:59 pm

+++ Frontend/src/pages/profile/ProfilePage.jsx
@@ -2,6 +2,7 @@ import { useEffect, useState } from 'react';
 import { getProfile, deleteProfile } from '../../Services/profileServices.js';
 import { User, Mail, Phone, Briefcase, Store, Calendar, Edit2, Trash2 } from "lucide-react";
 import { useNavigate } from 'react-router-dom';
+import toast from 'react-hot-toast';
 
 const ProfilePage = () => {
   const navigate = useNavigate();
@@ -19,15 +20,22 @@ const ProfilePage = () => {
 
   const handleDelete = async () => {
     try {
-      const deleted = await deleteProfile();
-      if(deleted){
-        console.log("Profile deleted seccessfully")
-        sessionStorage.clear();
-        navigate('/login')
+      if(confirm("Are you sure you want to delete your account? This will remove all your data.")){
+        const deleted = await deleteProfile();
+        if (deleted) {
+          console.log("Profile deleted seccessfully")
+          sessionStorage.clear();
+          navigate('/login')
+        }
       }
     } catch (err) {
+      toast.error('Failed to fetch profile!')
       console.error('Error fetching profile:', err);
     }
+  }
+
+  const handleEdit = async () => {
+    navigate("/profile/edit")
   }
 
   useEffect(() => {
@@ -37,10 +45,9 @@ const ProfilePage = () => {
   return (
     <div className="p-6 font-sans bg-gray-50">
       <div className="max-w-lg mx-auto bg-white rounded-2xl shadow-xl overflow-hidden">
-        {/* Header */}
         <div className="relative bg-gradient-to-r from-indigo-500 to-purple-600 p-6 text-center text-white">
-          <div className='absolute top-25 right-4 cursor-pointer text-indigo-600 bg-indigo-100 p-2 rounded-full'
-            onClick={() => navigate('/profile/edit')}>
+          <div className='absolute z-50 top-25 right-4 cursor-pointer text-indigo-600 bg-indigo-100 p-2 rounded-full'
+            onClick={() => handleEdit()}>
             {<Edit2 size={18} />}
           </div>
           <div className='absolute top-4 right-4 cursor-pointer text-indigo-600 bg-indigo-100 p-2 rounded-full'
@@ -49,13 +56,11 @@ const ProfilePage = () => {
           </div>
           <div className="w-fit mx-auto mb-2 p-2 rounded-full bg-white flex items-center justify-center font-bold text-2xl shadow-lg">
             <h1 className='text-black Logo-font'>{profile.shopname}</h1>
-            {/* <img src={Dukaan_Digital} alt="Dukaan_Digital" /> */}
           </div>
           <h2 className="text-xl font-semibold">{profile.name}</h2>
           <p className="text-sm opacity-80">{profile.role}</p>
         </div>
 
-        {/* Details */}
         <div className="p-6 space-y-4">
           <ProfileDetail icon={<Mail size={18} />} label="Email" value={profile.email} />
           <ProfileDetail icon={<Phone size={18} />} label="Phone" value={profile.phone} />
